+++ src/numberGame.css
@@ -1,0 +1,56 @@
+html {
+    background-image: url(/assets/spiderplant.jpg);
+    background-repeat: repeat;
+    border: 3px solid #636D52;
+    border-radius: 6px;
+    font-family: Arial, Helvetica, sans-serif;
+}   
+
+nav {
+    display: flex;
+    justify-content: space-around;
+    text-decoration: #636D52
+}
+
+body {
+    background: #C4C3B3;
+    padding: 6px;
+    border-radius: 6px;
+
+}
+
+Button:hover {
+    background: #C4C3B3; 
+}
+    
+Button:active {
+background: url(/assets/spiderplant.jpg); 
+}
+        
+.boxes {
+    width: 100vw;
+    display: flex;
+    justify-content: space-evenly;
+}
+        
+.feedback {
+    width: 100vw;
+    display: flex;
+    justify-content: space-evenly;
+}   
+
+.boxes *, .feedback * {
+   height: 50px;
+   width: 50px;
+   border: solid #636D52 3px;
+   border-radius: 6px;
+   margin-top: 30px;
+   margin-bottom: 40px;
+}
+
+#guessButton {
+    height: 60px;
+    width: 60px;
+}
+
+
